+++ src/Anixe.TransactionSteps/StepStat.cs
@@ -3,7 +3,7 @@ namespace Anixe.TransactionSteps
   public class StepStat
   {
     public string Name;
-    public int TimeTaken;
+    public double TimeTaken;
     public int ProcessedItemsCount;
 
     public static StepStat CreateFromStep(IStep step)
